examples/invoice-parser/requirements.txt
indexify-extractor-sdk==0.0.2
jupyterlab
Pillow
matplotlib